requirements.txt
flask
flask-cors
replicate
rembg
pillow
requests